Update tekton lib funcs to latest slsa v1.0 format
This commit updates several helper functions in the tekton package, to
accomodate for the latest format of a slsa v1.0 specification.

One of the main changes in the latest format is that task results are
now grouped in a byproducts section in the attestation, instead of being
an attribute of the task object.

Another change is that the task params are defined as templates (e.g:
$(params.hermetic), and need to be resolved by looking at the pipeline
params or at the byproducts section (in the case of task params coming
from other tasks results).

Because of these significant changes in the slsav1 format, several tests
across the whole repo had to be updated.

Note: this PR depends on https://github.com/tektoncd/chains/pull/1461 to
correctly resolve param templates from the byproducts section.

Ref: https://issues.redhat.com/browse/EC-1516